refactor: move old code from prod to tests
