+++ tests/test_tags.py
@@ -1,7 +1,7 @@
 import pytest
 from hypothesis import assume, given
 from hypothesis import strategies as st
-from repositories.tags import suggested_friends, suggested_friends_old
+from repositories.tags import suggested_friends
 
 
 @pytest.mark.asyncio()
@@ -23,3 +23,15 @@ async def test_suggested_friends_basic_2() -> None:
 async def test_suggested_friends_equal(xs: list[tuple[int, str]], amt: int) -> None:
     assume(amt > 0)
     assert await suggested_friends(xs, amt) == await suggested_friends_old(xs, amt)
+
+
+async def suggested_friends_old(result: list[tuple[int, str]], amt: int) -> dict[int, list[str]]:
+    # Group the results by user_id
+    suggestions: dict[int, list[str]] = {}
+    for suggested_user_id, tag in result:
+        if suggested_user_id not in suggestions:
+            suggestions[suggested_user_id] = []
+        suggestions[suggested_user_id].append(tag)
+
+    # Limit to top {amt} users with most common tags
+    return dict(sorted(suggestions.items(), key=lambda x: len(x[1]), reverse=True)[:amt])
